# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinic/portal/appointments/{id}:
    get:
      operationId: getPortalAppointment
      tags: [clinic]
      summary: Appointment detail for the patient or responsible professional
      parameters:
        - $ref: "#/components/parameters/RecordId"
      responses:
        "200":
          description: Appointment detail
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /clinic/portal/appointments/{id}/actions:
    post:
      operationId: performPortalAppointmentAction
      tags: [clinic]
      summary: Confirm, cancel, or request rescheduling
      parameters:
        - $ref: "#/components/parameters/RecordId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortalAppointmentActionInput"
      responses:
        "200":
          description: Idempotent action result
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  parameters:
...
    RecordId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    PortalAppointmentActionInput:
      type: object
      required: [action, confirmed, idempotencyKey]
      properties:
        action: { type: string, enum: [confirm_presence, request_reschedule, request_cancel] }
        confirmed: { type: boolean, enum: [true] }
        idempotencyKey: { type: string }
        requestedSlotId: { type: string }
...